fix type  card_id and user to int64

+++ pkg/merchantorder/request.go
@@ -37,5 +37,5 @@ type ItemRequest struct {
 
 // CollectorRequest represents seller information.
 type CollectorRequest struct {
-	ID int `json:"id,omitempty"`
+	ID int64 `json:"id,omitempty"`
 }
